+++ gulpfile.js/tasks/esbuild.mjs
@@ -1,14 +1,19 @@
 import DefaultRegistry from "undertaker-registry";
-import { createGulpEsbuild } from "gulp-esbuild";
+import esbuild from "gulp-esbuild";
 import projectPath from "../lib/projectPath.mjs";
 import debug from "gulp-debug";
 import logger from "gulplog";
 
 /** @typedef {import("@types/gulp")} Undertaker */
 export class ESBuildRegistry extends DefaultRegistry {
-  constructor(config, pathConfig, mode) {
+  constructor(config, pathConfig, mode, verbose) {
     super();
-    const modulePathConfig = pathConfig.esm ?? pathConfig.esbuild;
+    if (pathConfig.esbuild) {
+      logger.warn(
+        "`pathConfig.esbuild` is not supported. Use `pathConfig.esm` instead."
+      );
+    }
+    const modulePathConfig = pathConfig.esm;
     this.config = config;
     this.paths = {
       src: projectPath(
@@ -21,6 +26,10 @@ export class ESBuildRegistry extends DefaultRegistry {
       dest: projectPath(pathConfig.dest, modulePathConfig?.dest ?? "")
     };
     this.mode = mode;
+    this.verbose = verbose;
+    if (this.verbose) {
+      this.config.options.logLevel = "debug";
+    }
   }
 
   /**
@@ -29,16 +38,10 @@ export class ESBuildRegistry extends DefaultRegistry {
   init({ task, src, dest }) {
     if (!this.config) return;
 
-    const esbuild = createGulpEsbuild();
-    const esbuildInc = createGulpEsbuild({ incremental: true });
     task("esbuild", () =>
       src(this.paths.src)
         .pipe(debug({ title: "esbuild:", logger: logger.debug }))
-        .pipe(
-          this.mode.production()
-            ? esbuild(this.config.options)
-            : esbuildInc(this.config.options)
-        )
+        .pipe(esbuild(this.config.options))
         .pipe(dest(this.paths.dest))
     );
   }